requirements.txt
langchain==0.1.4
langchain_community==0.0.20
openai==1.12.0
pinecone-client==3.0.2
python-dotenv==1.0.1
sentence_transformers==2.3.1
torch==2.2.0
tiktoken==0.6.0
unstructured
ctransformers==0.2.27
langchain-openai==0.0.6
flask==3.0.2
PyMuPDF
fpdf
